[CC] moved square_and_smooth to loudness_zwicker folder
