+++ mosqito/functions/loudness_zwicker/calc_third_octave_levels.py
@@ -9,7 +9,7 @@ import numpy as np
 from scipy import signal
 
 # Local application imports
-from mosqito.functions.oct3filter.square_and_smooth import square_and_smooth
+from mosqito.functions.loudness_zwicker.square_and_smooth import square_and_smooth
 
 
 def calc_third_octave_levels(sig, fs):
